+++ packages/stacks-docs/plugins/header.js
@@ -1,5 +1,5 @@
 
-const { Icons } = require("@stackoverflow/stacks-icons");
+const { Icons } = require("@stackoverflow/stacks-icons-legacy");
 
 module.exports = {
   configFunction(eleventyConfig) {